+++ src/cli/middleware.ts
@@ -7,36 +7,43 @@ import { setPackageVersion } from '../helpers/set.js'
 import { InferredOptionTypes, MiddlewareFunction } from 'yargs'
 import { CommanderPackage } from '../translator/commander.types.js'
 import cmdr from '../translator/commander.js'
-import { options } from './swpx/cli.js'
+import { options as swpmOptions } from './swpm/cli.js'
+import { options as swpxOptions } from './swpx/cli.js'
+import { PackageManagerList } from '../packages/packages.types.js'
 
-const middleware: MiddlewareFunction<InferredOptionTypes<typeof options>> = async (yargs) => {
+type Props = InferredOptionTypes<typeof swpmOptions> | InferredOptionTypes<typeof swpxOptions>
+const middleware: MiddlewareFunction<Props> = async (yargs) => {
   cmdr.args = argv.slice(2)
 
-  if (yargs?.debug) {
+  if ('debug' in yargs) {
     cleanFlag({ yargs, cmdr, flag: '--debug' })
     cleanFlag({ yargs, cmdr, flag: '-d' })
   }
 
-  if (yargs?.use) {
+  if ('use' in yargs) {
     cleanFlag({ yargs, cmdr, flag: '--use' })
     cleanFlag({ yargs, cmdr, flag: '-u' })
     cmdr.cmd = yargs.use
     await setPackageVersion(yargs.use!)
   }
 
-  if (yargs?.test) {
+  if ('pin' in yargs && yargs?.pin) {
+    cmdr.cmd = yargs.pin as PackageManagerList
+  }
+
+  if ('test' in yargs) {
     cleanFlag({ yargs, cmdr, flag: '--test' })
     cleanFlag({ yargs, cmdr, flag: '-t' })
     cmdr.cmd = yargs.test
   }
 
-  if (yargs?.mute) {
+  if ('mute' in yargs) {
     cleanFlag({ yargs, cmdr, flag: '--mute' })
     cleanFlag({ yargs, cmdr, flag: '-m' })
   }
 
-  if (!(cmdr?.cmd) || yargs?.info) {
-    const { origin, cmd } = await getCurrentPackageManager()
+  if (!cmdr?.cmd || yargs?.info) {
+    const { origin, cmd } = await getCurrentPackageManager() || {}
     cmdr.origin = origin
     cmdr.cmd = cmd
   }
@@ -44,14 +51,14 @@ const middleware: MiddlewareFunction<InferredOptionTypes<typeof options>> = asyn
   if (cmdr?.cmd) {
     cmdr.volta = await detectVoltaPin(cmdr as CommanderPackage) ?? false
     cmdr.config = await getPackageConfiguration(cmdr as CommanderPackage)
-  }
 
-  if (yargs?.global) {
-    translateArgs({ yargs, cmdr, flag: '--global', alias: '-g' })
-  }
+    if ('global' in yargs) {
+      translateArgs({ yargs, cmdr, flag: '--global', alias: '-g' })
+    }
 
-  if (yargs._.length) {
-    translateCommand({ yargs, cmdr })
+    if (yargs._.length) {
+      translateCommand({ yargs, cmdr })
+    }
   }
 }