+++ src/cli/options/pin.ts
@@ -4,7 +4,7 @@ const pin = {
   alias: 'p',
   describe: 'pin a package manager',
   choices: availablePackages(),
-  conflicts: ['use']
+  conflicts: ['use', 'unpin']
 } as const
 
 export default pin