+++ src/cli/options/use.ts
@@ -4,7 +4,7 @@ const use = {
   alias: 'u',
   describe: 'use a package manager',
   choices: availablePackages(),
-  conflicts: ['pin']
+  conflicts: ['pin', 'unpin']
 } as const
 
 export default use