+++ src/cli/swpm.ts
@@ -2,32 +2,71 @@
 
 import yargs from './swpm/config.js'
 
+import prompts from 'prompts'
+import chalk from 'chalk'
+import { stripIndent } from 'common-tags'
+import { exit } from 'node:process'
+
+import { autoUpdate } from '../libs/autoUpdate.js'
+
 import { pinPackageManager } from '../flags/pin.js'
-import { showPackageInformation } from '../flags/info.js'
+import { unpinPackageManager } from '../flags/unpin.js'
+import { showNoPackageDetected, showPackageInformation } from '../flags/info.js'
 import { showCommandAlias } from '../flags/alias.js'
 import { testCommand } from '../flags/test.js'
 
-import { autoUpdate } from '../libs/autoUpdate.js'
 import { showCommand, runCommand } from '../helpers/cmds.js'
 import { setPackageVersion } from '../helpers/set.js'
-import cmdr from '../translator/commander.js'
-import { CommanderPackage } from '../translator/commander.types.js'
 import { debug } from '../helpers/debug.js'
+import { commandVerification } from '../helpers/get.js'
+
+import cmdr from '../translator/commander.js'
 
 if (yargs.debug) {
   debug(yargs)
   debug(cmdr)
 }
 
-await autoUpdate(cmdr)
+if (cmdr?.cmd) {
+  await autoUpdate(cmdr)
+}
 
 if (yargs?.pin) {
-  cmdr.cmd = yargs.pin!
-  await setPackageVersion(cmdr.cmd!)
-
   const { cmd, config } = cmdr
+
+  const isInstalled = !!cmd && await commandVerification(cmd)
+
+  if (!isInstalled) {
+    const color = config?.color ?? chalk.reset()
+
+    const response = await prompts({
+      type: 'confirm',
+      name: 'value',
+      message: stripIndent`
+        ${chalk.hex(color).bold(cmd)} is not installed.
+        Do you still want to continue?
+      `,
+      initial: false
+    })
+
+    if (!response.value) {
+      exit(1)
+    }
+  }
+
+  if (cmd) {
+    await setPackageVersion(cmd)
+  }
+
   if (cmd && config) {
     await pinPackageManager({ cmd, config })
+  }
+}
+
+if (yargs?.unpin) {
+  const { config } = cmdr
+  if (config) {
+    await unpinPackageManager({ config })
   }
 }
 
@@ -35,24 +74,19 @@ if (yargs?.test) {
   testCommand(cmdr)
 }
 
-if (yargs?.info && cmdr?.cmd && cmdr?.config && cmdr?.origin && cmdr?.volta) {
-  const cmdrInfo: Required<CommanderPackage> = {
-    cmd: cmdr.cmd,
-    args: cmdr.args,
-    origin: cmdr.origin,
-    config: cmdr.config,
-    volta: cmdr.volta
-  }
-  await showPackageInformation(cmdrInfo)
+if (yargs?.info) {
+  await showPackageInformation(cmdr)
 }
 
 if (yargs?.alias) {
   await showCommandAlias()
 }
 
-if (cmdr?.cmd) {
-  if (!yargs?.mute) {
-    showCommand(cmdr)
-  }
-  await runCommand(cmdr)
+if (!cmdr?.cmd) {
+  showNoPackageDetected()
 }
+
+if (!yargs?.mute) {
+  showCommand(cmdr)
+}
+await runCommand(cmdr)