+++ src/cli/swpm/cli.ts
@@ -9,6 +9,7 @@ import clean from './commands/clean.js'
 import debug from '../options/debug.js'
 import use from '../options/use.js'
 import pin from '../options/pin.js'
+import unpin from '../options/unpin.js'
 import test from '../options/test.js'
 import mute from '../options/mute.js'
 import alias from '../options/alias.js'
@@ -28,6 +29,7 @@ const options = {
   debug,
   use,
   pin,
+  unpin,
   test,
   mute,
   alias,