+++ src/cli/swpm/commands/install.ts
@@ -35,15 +35,17 @@ const install: CommandModule<Record<string, unknown>, OptionsProps> = {
       } as const),
 
   handler: async (yargs) => {
-    if (yargs?.['package-lock']) {
+    if (!cmdr?.cmd) return
+
+    if ('package-lock' in yargs) {
       translateArgs({ yargs, cmdr, flag: '--package-lock', alias: '-P' })
     }
 
-    if (yargs?.frozen) {
+    if ('frozen' in yargs) {
       translateArgs({ yargs, cmdr, flag: '--frozen', alias: '-F' })
     }
 
-    if (yargs.FLAGS || yargs?.global) {
+    if (('FLAGS' in yargs) || ('global' in yargs)) {
       const args = ['add', ...cmdr.args.slice(1)]
       const command = chalk.blue.bold(`swpm ${args.join(' ')}`)
 
@@ -54,14 +56,11 @@ const install: CommandModule<Record<string, unknown>, OptionsProps> = {
       const response = await prompts({
         type: 'confirm',
         name: 'value',
-        message: `Do you want to re-run as ${command}`,
+        message: `Do you want to re-run it as ${command}`,
         initial: true
       })
 
       if (!response.value) {
-        console.error(
-          stripIndents`Re-run as ${command}`
-        )
         exit(1)
       }